fix: eliminate client-side retries and back button issues
- Moved event processing from separate components to form submission handlers
- Removed EventsFromUrl and EventsFromRawText components completely
- Process events directly in AddEvent component on form submit
- Navigate to upcoming page only after starting workflow
- Added processing state to text and URL form buttons
- No more URL params that trigger processing on mount/remount

This prevents duplicate event creation when using back button or navigating back to /new page.
